#80-fix: comment js에서 누락된 변수 선언을 추가 해주었다.
