+++ geulkkoli/src/main/java/com/geulkkoli/application/social/service/CustomOauth2UserService.java
@@ -8,8 +8,8 @@ import com.geulkkoli.application.user.ProviderUser;
 import com.geulkkoli.application.user.UserModelDto;
 import com.geulkkoli.application.social.util.DelegateOAuth2RequestConverter;
 import com.geulkkoli.application.social.util.ProviderUserRequest;
-import com.geulkkoli.domain.social.SocialInfoFindService;
-import com.geulkkoli.domain.social.SocialInfoService;
+import com.geulkkoli.domain.social.service.SocialInfoFindService;
+import com.geulkkoli.domain.social.service.SocialInfoService;
 import com.geulkkoli.domain.user.User;
 import com.geulkkoli.domain.user.service.UserFindService;
 import lombok.extern.slf4j.Slf4j;
