+++ geulkkoli/src/main/java/com/geulkkoli/domain/admin/service/AdminServiceImpl.java
@@ -1,17 +1,17 @@
 package com.geulkkoli.domain.admin.service;
 
 import com.geulkkoli.domain.admin.AccountLock;
-import com.geulkkoli.domain.admin.AccountLockRepository;
-import com.geulkkoli.domain.admin.ReportRepository;
 import com.geulkkoli.domain.hashtag.HashTag;
 import com.geulkkoli.domain.post.Post;
 import com.geulkkoli.domain.post.PostRepository;
 import com.geulkkoli.domain.posthashtag.PostHashTag;
-import com.geulkkoli.domain.posthashtag.PostHashTagService;
+import com.geulkkoli.domain.posthashtag.service.PostHashTagService;
 import com.geulkkoli.domain.topic.Topic;
 import com.geulkkoli.domain.topic.TopicRepository;
 import com.geulkkoli.domain.user.User;
 import com.geulkkoli.domain.user.UserRepository;
+import com.geulkkoli.domain.admin.AccountLockRepository;
+import com.geulkkoli.domain.admin.ReportRepository;
 import com.geulkkoli.web.admin.DailyTopicDto;
 import com.geulkkoli.web.admin.ReportDto;
 import com.geulkkoli.web.post.dto.AddDTO;
