+++ geulkkoli/src/main/java/com/geulkkoli/domain/social/service/SocialInfoService.java
@@ -1,5 +1,7 @@
-package com.geulkkoli.domain.social;
+package com.geulkkoli.domain.social.service;
 
+import com.geulkkoli.domain.social.SocialInfo;
+import com.geulkkoli.domain.social.SocialInfoRepository;
 import com.geulkkoli.web.social.SocialInfoDto;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
