+++ geulkkoli/src/main/resources/application.yml
@@ -7,13 +7,14 @@ spring:
     multipart:
       max-file-size: 10MB
       max-request-size: 100MB
+      location:
 
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
     url: ${spring.datasource.url}
     username: ${spring.datasource.username}
     password: ${spring.datasource.password}
- # creat- drop -> update로 수정 (테이블이 없으면 생성하고 있으면 수정)
+  # creat- drop -> update로 수정 (테이블이 없으면 생성하고 있으면 수정)
   jpa:
     hibernate:
       ddl-auto: update
@@ -116,4 +117,4 @@ server:
     include-message: never
 
 comm:
-  uploadPath: C:/geulkkoliFile/
+  uploadPath: C:/Users/82109/Desktop/Project/Project/src/main/resources/static/upload/
