+++ geulkkoli/src/test/java/com/geulkkoli/domain/comment/CommentServiceTest.java
@@ -1,12 +1,11 @@
 package com.geulkkoli.domain.comment;
 
+import com.geulkkoli.domain.comment.service.CommentsService;
 import com.geulkkoli.domain.post.Post;
 import com.geulkkoli.domain.user.User;
 import com.geulkkoli.web.comment.dto.CommentBodyDTO;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentMatcher;
-import org.mockito.ArgumentMatchers;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
