+++ geulkkoli/src/test/java/com/geulkkoli/domain/post/service/PostFindServiceTest.java
@@ -6,16 +6,16 @@ import com.geulkkoli.web.post.dto.AddDTO;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
+@WebAppConfiguration
 class PostFindServiceTest {
     @Autowired
     UserRepository userRepository;
@@ -34,9 +34,7 @@ class PostFindServiceTest {
                 .nickName("nickName")
                 .build();
 
-        LocalDate signUpDate = LocalDate.of(2022, 2, 2);
-
-        userRepository.save(user);
+        User save = userRepository.save(user);
 
         AddDTO addDTO1 = AddDTO.builder()
                 .title("달력")
@@ -44,19 +42,19 @@ class PostFindServiceTest {
                 .nickName(user.getNickName())
                 .build();
         LocalDateTime createAt1 = LocalDateTime.of(2022, 3, 7, 1, 1);
-        user.writePost(addDTO1, createAt1);
+        save.writePost(addDTO1);
 
         AddDTO addDTO2 = AddDTO.builder()
                 .title("달력")
                 .postBody("요호")
-                .nickName(user.getNickName())
+                .nickName(save.getNickName())
                 .build();
         LocalDateTime createAt2 = LocalDateTime.of(2023, 4, 8, 12, 1);
-        user.writePost(addDTO2, createAt2);
+        save.writePost(addDTO2);
 
-        List<String> createdAts = postFindService.getCreatedAts(user);
+        List<String> createdAts = postFindService.getCreatedAts(save);
 
-        assertThat(createdAts.size()).isEqualTo(2);
+        assertThat(createdAts).hasSize(2);
         assertThat(createdAts).contains(String.valueOf(createAt1.toLocalDate()));
         assertThat(createdAts).contains(String.valueOf(createAt2.toLocalDate()));
